Refactored.
Moved models into its own file in the config package and made main source of origin.  Next up is moving agent settings away from main into the config file.

+++ src/syn_grid/core/resources/base_resource.py
@@ -27,7 +27,7 @@ class BaseResource(ABC):
 
     def __init__(
         self,
-        reward: int,
+        reward: float,
         cool_down: int,
         meta: ResourceMeta,
     ):
